Corrections for changes in the compiler and math and a few features.
Added the 'set to 0.0, if less than 2W'. Also found that with SDCC 2.6.0 math handling is incorrect. Not sure if it was a bug in my code or the 2.3 int and long were unsigned. But I did need to migrate some of the values to unsigned long as I am well past 65535. Also added an ID string that gets spit out on reset, version information.